+++ tests/ui/test_ui_Scanner.m
@@ -1,0 +1,27 @@
+[cDirThis, cName, cExt] = fileparts(mfilename('fullpath'));
+
+% bl12014 src
+cDirBl12014 = fullfile(cDirThis, '..', '..', 'src');
+addpath(genpath(cDirBl12014));
+
+% dependencies
+cDirVendor = fullfile(cDirThis, '..', '..', 'vendor');
+
+addpath(genpath(fullfile(cDirVendor, 'github', 'cnanders', 'matlab-instrument-control', 'src')));
+addpath(genpath(fullfile(cDirVendor, 'github', 'cnanders', 'matlab-quasar', 'src')));
+addpath(genpath(fullfile(cDirVendor, 'github', 'cnanders', 'matlab-pupil-fill-generator', 'src')));
+addpath(genpath(fullfile(cDirVendor, 'github', 'cnanders', 'matlab-npoint-lc400-ui', 'src')));
+
+purge
+
+clock = mic.Clock('Master');
+
+ui = bl12014.ui.Scanner(...
+    'clock', clock ...
+);
+
+ui.build();
+
+
+ 
+