+++ tests/ui/test_ui_Scanner.m
@@ -11,7 +11,7 @@ addpath(genpath(fullfile(cDirVendor, 'github', 'cnanders', 'matlab-instrument-co
 addpath(genpath(fullfile(cDirVendor, 'github', 'cnanders', 'matlab-quasar', 'src')));
 addpath(genpath(fullfile(cDirVendor, 'github', 'cnanders', 'matlab-pupil-fill-generator', 'src')));
 addpath(genpath(fullfile(cDirVendor, 'github', 'cnanders', 'matlab-npoint-lc400-ui', 'src')));
-
+addpath(genpath(fullfile(cDirVendor, 'github', 'cnanders', 'matlab-npoint-lc400', 'src')));
 purge
 
 clock = mic.Clock('Master');